requirements.txt
beautifulsoup4>=4.12
requests>=2.31
pandas>=2.2
python-dotenv>=1.0
typing-extensions>=4.5; python_version<3.9
